refactor(store): 重构了store的目录
为了后面可以扩展更多store，先把store分配到一个新的store包中
