+++ captcha.go
@@ -19,6 +19,7 @@ package base64Captcha
 import (
 	"encoding/base64"
 	"fmt"
+	"github.com/kpaas-io/base64Captcha/store"
 	"io"
 	"log"
 	"os"
@@ -36,27 +37,25 @@ var (
 	Expiration = 10 * time.Minute
 	// globalStore is a shared storage for captchas, generated by New function.
 	// 默认内存储存
-	globalStore = NewMemoryStore(GCLimitNumber, Expiration)
-	//thisPackageDirPath current package path.
-	thisPackageDirPath = ""
+	globalStore = store.NewMemoryStore(GCLimitNumber, Expiration)
 )
 
 // SetCustomStore sets custom storage for captchas, replacing the default
 // memory store. This function must be called before generating any captchas.
-func SetCustomStore(s Store) {
+func SetCustomStore(s store.Store) {
 	globalStore = s
 }
 
-//CaptchaInterface captcha interface for captcha engine to to write staff
+// CaptchaInterface captcha interface for captcha engine to to write staff
 type CaptchaInterface interface {
-	//BinaryEncodeing covert to bytes
+	// BinaryEncodeing covert to bytes
 	BinaryEncodeing() []byte
-	//WriteTo output captcha entity
+	// WriteTo output captcha entity
 	WriteTo(w io.Writer) (n int64, err error)
 }
 
-//CaptchaWriteToBase64Encoding converts captcha to base64 encoding string.
-//mimeType is one of "audio/wav" "image/png".
+// CaptchaWriteToBase64Encoding converts captcha to base64 encoding string.
+// mimeType is one of "audio/wav" "image/png".
 func CaptchaWriteToBase64Encoding(cap CaptchaInterface) string {
 	binaryData := cap.BinaryEncodeing()
 	var mimeType string
@@ -68,8 +67,8 @@ func CaptchaWriteToBase64Encoding(cap CaptchaInterface) string {
 	return fmt.Sprintf("data:%s;base64,%s", mimeType, base64.StdEncoding.EncodeToString(binaryData))
 }
 
-//CaptchaWriteToFile output captcha to file.
-//fileExt is one of "png","wav"
+// CaptchaWriteToFile output captcha to file.
+// fileExt is one of "png","wav"
 func CaptchaWriteToFile(cap CaptchaInterface, outputDir, fileName, fileExt string) error {
 	filePath := filepath.Join(outputDir, fileName+"."+fileExt)
 	file, err := os.Create(filePath)
@@ -82,15 +81,15 @@ func CaptchaWriteToFile(cap CaptchaInterface, outputDir, fileName, fileExt strin
 	return err
 }
 
-//CaptchaItem captcha basic information.
+// CaptchaItem captcha basic information.
 type CaptchaItem struct {
-	//Content captcha entity content.
+	// Content captcha entity content.
 	Content string
-	//VerifyValue captcha verify value.
+	// VerifyValue captcha verify value.
 	VerifyValue string
-	//ImageWidth image width pixel.
+	// ImageWidth image width pixel.
 	ImageWidth int
-	//ImageHeight image height pixel.
+	// ImageHeight image height pixel.
 	ImageHeight int
 }
 
@@ -113,56 +112,56 @@ func VerifyCaptchaAndIsClear(identifier, verifyValue string, isClear bool) bool 
 	return strings.ToLower(storeValue) == strings.ToLower(verifyValue)
 }
 
-//GenerateCaptcha create captcha by config struct and id.
-//idkey can be an empty string, base64 will create a unique id four you.
-//if idKey is a empty string, the package will generate a random unique identifier for you.
-//configuration struct should be one of those struct ConfigAudio, ConfigCharacter, ConfigDigit.
+// GenerateCaptcha create captcha by config struct and id.
+// idkey can be an empty string, base64 will create a unique id four you.
+// if idKey is a empty string, the package will generate a random unique identifier for you.
+// configuration struct should be one of those struct ConfigAudio, ConfigCharacter, ConfigDigit.
 //
-//Example Code
-//	//config struct for digits
-//	var configD = base64Captcha.ConfigDigit{
-//		Height:     80,
-//		Width:      240,
-//		MaxSkew:    0.7,
-//		DotCount:   80,
-//		CaptchaLen: 5,
-//	}
-//	//config struct for audio
-//	var configA = base64Captcha.ConfigAudio{
-//		CaptchaLen: 6,
-//		Language:   "zh",
-//	}
-//	//config struct for Character
-//	var configC = base64Captcha.ConfigCharacter{
-//		Height:             60,
-//		Width:              240,
-//		//const CaptchaModeNumber:数字,CaptchaModeAlphabet:字母,CaptchaModeArithmetic:算术,CaptchaModeNumberAlphabet:数字字母混合.
-//		Mode:               base64Captcha.CaptchaModeNumber,
-//		ComplexOfNoiseText: base64Captcha.CaptchaComplexLower,
-//		ComplexOfNoiseDot:  base64Captcha.CaptchaComplexLower,
-//		IsUseSimpleFont:    true,
-//		IsShowHollowLine:   false,
-//		IsShowNoiseDot:     false,
-//		IsShowNoiseText:    false,
-//		IsShowSlimeLine:    false,
-//		IsShowSineLine:     false,
-//		CaptchaLen:         6,
-//	}
-//	//create a audio captcha.
-//	//GenerateCaptcha first parameter is empty string,so the package will generate a random uuid for you.
-//	idKeyA,capA := base64Captcha.GenerateCaptcha("",configA)
-//	//write to base64 string.
-//	//GenerateCaptcha first parameter is empty string,so the package will generate a random uuid for you.
-//	base64stringA := base64Captcha.CaptchaWriteToBase64Encoding(capA)
-//	//create a characters captcha.
-//	//GenerateCaptcha first parameter is empty string,so the package will generate a random uuid for you.
-//	idKeyC,capC := base64Captcha.GenerateCaptcha("",configC)
-//	//write to base64 string.
-//	base64stringC := base64Captcha.CaptchaWriteToBase64Encoding(capC)
-//	//create a digits captcha.
-//	idKeyD,capD := base64Captcha.GenerateCaptcha("",configD)
-//	//write to base64 string.
-//	base64stringD := base64Captcha.CaptchaWriteToBase64Encoding(capD)
+// Example Code
+// 	//config struct for digits
+// 	var configD = base64Captcha.ConfigDigit{
+// 		Height:     80,
+// 		Width:      240,
+// 		MaxSkew:    0.7,
+// 		DotCount:   80,
+// 		CaptchaLen: 5,
+// 	}
+// 	//config struct for audio
+// 	var configA = base64Captcha.ConfigAudio{
+// 		CaptchaLen: 6,
+// 		Language:   "zh",
+// 	}
+// 	//config struct for Character
+// 	var configC = base64Captcha.ConfigCharacter{
+// 		Height:             60,
+// 		Width:              240,
+// 		//const CaptchaModeNumber:数字,CaptchaModeAlphabet:字母,CaptchaModeArithmetic:算术,CaptchaModeNumberAlphabet:数字字母混合.
+// 		Mode:               base64Captcha.CaptchaModeNumber,
+// 		ComplexOfNoiseText: base64Captcha.CaptchaComplexLower,
+// 		ComplexOfNoiseDot:  base64Captcha.CaptchaComplexLower,
+// 		IsUseSimpleFont:    true,
+// 		IsShowHollowLine:   false,
+// 		IsShowNoiseDot:     false,
+// 		IsShowNoiseText:    false,
+// 		IsShowSlimeLine:    false,
+// 		IsShowSineLine:     false,
+// 		CaptchaLen:         6,
+// 	}
+// 	//create a audio captcha.
+// 	//GenerateCaptcha first parameter is empty string,so the package will generate a random uuid for you.
+// 	idKeyA,capA := base64Captcha.GenerateCaptcha("",configA)
+// 	//write to base64 string.
+// 	//GenerateCaptcha first parameter is empty string,so the package will generate a random uuid for you.
+// 	base64stringA := base64Captcha.CaptchaWriteToBase64Encoding(capA)
+// 	//create a characters captcha.
+// 	//GenerateCaptcha first parameter is empty string,so the package will generate a random uuid for you.
+// 	idKeyC,capC := base64Captcha.GenerateCaptcha("",configC)
+// 	//write to base64 string.
+// 	base64stringC := base64Captcha.CaptchaWriteToBase64Encoding(capC)
+// 	//create a digits captcha.
+// 	idKeyD,capD := base64Captcha.GenerateCaptcha("",configD)
+// 	//write to base64 string.
+// 	base64stringD := base64Captcha.CaptchaWriteToBase64Encoding(capD)
 func GenerateCaptcha(idKey string, configuration interface{}) (id string, captchaInstance CaptchaInterface) {
 	if idKey == "" {
 		idKey = randomId()
