+++ store/memory.go
@@ -12,29 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package base64Captcha
+package store
 
 import (
 	"container/list"
 	"sync"
 	"time"
 )
-
-// Store An object implementing Store interface can be registered with SetCustomStore
-// function to handle storage and retrieval of captcha ids and solutions for
-// them, replacing the default memory store.
-//
-// It is the responsibility of an object to delete expired and used captchas
-// when necessary (for example, the default memory store collects them in Set
-// method after the certain amount of captchas has been stored.)
-type Store interface {
-	// Set sets the digits for the captcha id.
-	Set(id string, value string)
-
-	// Get returns stored digits for the captcha id. Clear indicates
-	// whether the captcha must be deleted from the store.
-	Get(id string, clear bool) string
-}
 
 // expValue stores timestamp and id of captchas. It is used in the list inside
 // memoryStore for indexing generated captchas by timestamp to enable garbage
